worked around a skipping issue found in many webm players
Google Chrome, mpv, ffplay and many others are unable to skip back to the start of the audio in a webm. They are able to play from the start on initial load, but will be unable to jump back to the start. This is most obvious when the audio starts on drums.
this fix just subtracts 0.1 from the start calculation, meaning that the added audio will be mostly inaudible.

This problem did not affect Firefox, but the 0.1 seconds added will only be audible if the listener knows that it is there.